#5 Fixed some minor build issues

+++ P5OSPPB/kserver/kserver.c
@@ -210,7 +210,7 @@ void post_to_kern(unsigned int source, unsigned int command, unsigned int payloa
             break;
 
         case KS_GET_SHARED_PAGE:
-            passMessage(0, source, command, allocate_shared_page());
+            passMessage(0, source, command, (unsigned int)allocate_shared_page());
             break;
 
         default:
